backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
pydantic-settings
python-dotenv
cryptography
psycopg2-binary
email-validator
